feat: Add animated courses and experiences items with visibility detection
